Move boot probe augmentation helper into package
